refactor json tests to use mixin classes
