+++ astropy/cosmology/tests/test_flrw.py
@@ -14,21 +14,20 @@ import pytest
 
 import numpy as np
 
+import astropy.constants as const
 # LOCAL
 import astropy.cosmology.units as cu
 import astropy.units as u
-import astropy.constants as const
 from astropy.cosmology import (FLRW, FlatLambdaCDM, Flatw0waCDM, FlatwCDM,
                                LambdaCDM, Planck18, w0waCDM, w0wzCDM, wCDM, wpwaCDM)
 from astropy.cosmology.core import _COSMOLOGY_CLASSES
-from astropy.cosmology.flrw import (a_B_c2, critdens_const, ellipkinc,
-                                    H0units_to_invs, hyp2f1, quad)
+from astropy.cosmology.flrw import H0units_to_invs, a_B_c2, critdens_const, ellipkinc, hyp2f1, quad
 from astropy.cosmology.parameter import Parameter
 from astropy.utils.compat.optional_deps import HAS_SCIPY
 
 from .conftest import get_redshift_methods
-from .test_core import CosmologySubclassTest as CosmologyTest, valid_zs, invalid_zs
-from .test_core import FlatCosmologyMixinTest, ParameterTestMixin
+from .test_core import CosmologySubclassTest as CosmologyTest
+from .test_core import FlatCosmologyMixinTest, ParameterTestMixin, invalid_zs, valid_zs
 
 
 ##############################################################################
